+++ parser_library/src/context/ordinary_assembly/ordinary_assembly_context.h
@@ -103,6 +103,7 @@ public:
     const symbol* get_symbol_reference(context::id_index name) const;
 
     symbol* get_symbol(id_index name);
+    const symbol* get_symbol(id_index name) const;
 
     // gets section by name
     section* get_section(id_index name);